+++ crates/apollo-compiler/src/schema/validation.rs
@@ -11,7 +11,7 @@ pub(crate) fn validate_schema(
     errors: &mut Diagnostics,
     schema: &Schema,
 ) -> Vec<crate::ApolloDiagnostic> {
-    for (&file_id, source) in &schema.sources {
+    for (&file_id, source) in schema.sources.iter() {
         source.validate_parse_errors(errors, file_id)
     }
     for build_error in &schema.build_errors {
@@ -47,7 +47,7 @@ fn validate_build_error(errors: &mut Diagnostics, build_error: &BuildError) {
 fn compiler_validation(errors: &mut Diagnostics, schema: &Schema) -> Vec<crate::ApolloDiagnostic> {
     let mut compiler = crate::ApolloCompiler::new();
     let mut ids = Vec::new();
-    for (id, source) in &schema.sources {
+    for (id, source) in schema.sources.iter() {
         ids.push(*id);
         compiler.db.set_input(*id, source.into());
     }